feat: able to show timesheet

+++ server-core/src/main/java/io/onedev/server/infomanager/IssueInfoManager.java
@@ -6,8 +6,8 @@ import io.onedev.server.model.Issue;
 
 public interface IssueInfoManager {
 
-	Map<Integer, String> getDailyStates(Issue issue, Integer fromDay, Integer toDay);
+	Map<Long, String> getDailyStates(Issue issue, Long fromDay, Long toDay);
 	
-	Map<Integer, Integer> getDailySpentTimes(Issue issue, Integer fromDay, Integer toDay);
+	Map<Long, Integer> getDailySpentTimes(Issue issue, Long fromDay, Long toDay);
 	
 }
